- Added windows compatibiligy - Added py-lmd library to requirements.txt

+++ pipex.py
@@ -83,7 +83,8 @@ def batch_processor():
                     arg_start_index = arg_start_index + 1
                 arg_end_index = curr_command.index(end_char, arg_start_index + 1)
                 curr_data_folder = curr_command[arg_start_index:arg_end_index].strip()
-                shutil.copyfile(log_filename, curr_data_folder + '/' + os.path.basename(log_filename))
+                if os.path.exists(log_filename):
+                    shutil.copyfile(log_filename, curr_data_folder + '/' + os.path.basename(log_filename))
  
     batch_file.close()
